Add patched ESRI Shapefile nodes to repository
Luckily there are some nodes to import shapefiles into KNIME. Unfortunately they are buggy and do not work. This commit adds a patched fork of those nodes to the Cyface repository. This results also in a reorganization of the whole repository from a single project to a multi project variant. A future commit needs to provide a multi project POM as well.