Move Jenkinsfile to top level and change build dir
Since Jenkins can find the Jenkinsfile only on the top level this commit moves the Jenkinsfile which was designed for the Cyface nodes only to the root folder. Currently this file is only capable of building the Cyface nodes, so all the command must run inside the `de.cyface.knime` subfolder. The Jenkinsfile was updated to do just that. Hopefully this works.