Fix bug resetting input column for smooth node
This commit fixes a bug that was part of the configure methode of the smooth node. It had been resetting the input column all the time. To find this bug several checks have been included and a little logging was added to the SmoothingNodeModel. I also tried some things related to the bug in the MapMatchingNode. That stuff is going to disappear in the final implementation.